mz560: block hijacking via COPY

+++ pkg/util/fs_util.go
@@ -775,6 +775,10 @@ func CopySymlink(src, dest string, context FileContext) (bool, error) {
 		logrus.Debugf("%s found in .dockerignore, ignoring", src)
 		return true, nil
 	}
+	if CheckIgnoreList(dest) {
+		logrus.Debugf("Skipping copy for ignored path: %s", dest)
+		return true, nil
+	}
 	if FilepathExists(dest) {
 		if err := os.RemoveAll(dest); err != nil {
 			return false, err
@@ -794,6 +798,10 @@ func CopySymlink(src, dest string, context FileContext) (bool, error) {
 func CopyFile(src, dest string, context FileContext, uid, gid int64, chmod fs.FileMode, useDefaultChmod bool) (bool, error) {
 	if context.ExcludesFile(src) {
 		logrus.Debugf("%s found in .dockerignore, ignoring", src)
+		return true, nil
+	}
+	if CheckIgnoreList(dest) {
+		logrus.Debugf("Skipping copy for ignored path: %s", dest)
 		return true, nil
 	}
 	if src == dest {
